fix bug query order service

+++ product-service/_infra/14_export_order_detail.sql
@@ -4,7 +4,7 @@ CREATE TABLE `export_order_detail` (
                                        `variant_id` int DEFAULT NULL,
                                        `variant_name` varchar(500) DEFAULT NULL,
                                        `quantity` int DEFAULT NULL,
-                                       `price` double DEFAULT NULL,
+                                       `price` int DEFAULT NULL,
                                        PRIMARY KEY (`export_order_detail_id`),
                                        KEY `fk_export_orders_idx` (`export_order_id`),
                                        CONSTRAINT `fk_export_orders_export_order_detail` FOREIGN KEY (`export_order_id`) REFERENCES `export_orders` (`export_order_id`)
